Move the '//api/database' and '//api/user' package into '//api/connection'
