+++ api/connection/TypeDBClient.java
@@ -21,8 +21,8 @@
 
 package com.vaticle.typedb.client.api.connection;
 
-import com.vaticle.typedb.client.api.database.DatabaseManager;
-import com.vaticle.typedb.client.api.user.UserManager;
+import com.vaticle.typedb.client.api.connection.database.DatabaseManager;
+import com.vaticle.typedb.client.api.connection.user.UserManager;
 
 import javax.annotation.CheckReturnValue;
