+++ api/connection/user/User.java
@@ -19,7 +19,7 @@
  * under the License.
  */
 
-package com.vaticle.typedb.client.api.user;
+package com.vaticle.typedb.client.api.connection.user;
 
 import javax.annotation.CheckReturnValue;
