+++ connection/TypeDBDatabaseImpl.java
@@ -21,7 +21,7 @@
 
 package com.vaticle.typedb.client.connection;
 
-import com.vaticle.typedb.client.api.database.Database;
+import com.vaticle.typedb.client.api.connection.database.Database;
 import com.vaticle.typedb.client.common.rpc.TypeDBStub;
 
 import static com.vaticle.typedb.client.common.rpc.RequestBuilder.Core.Database.deleteReq;
