+++ connection/cluster/ClusterSession.java
@@ -24,7 +24,7 @@ package com.vaticle.typedb.client.connection.cluster;
 import com.vaticle.typedb.client.api.connection.TypeDBOptions;
 import com.vaticle.typedb.client.api.connection.TypeDBSession;
 import com.vaticle.typedb.client.api.connection.TypeDBTransaction;
-import com.vaticle.typedb.client.api.database.Database;
+import com.vaticle.typedb.client.api.connection.database.Database;
 import com.vaticle.typedb.client.connection.TypeDBSessionImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
