+++ connection/cluster/ClusterUser.java
@@ -21,7 +21,7 @@
 
 package com.vaticle.typedb.client.connection.cluster;
 
-import com.vaticle.typedb.client.api.user.User;
+import com.vaticle.typedb.client.api.connection.user.User;
 
 import static com.vaticle.typedb.client.common.rpc.RequestBuilder.Cluster.User.deleteReq;
 import static com.vaticle.typedb.client.connection.cluster.ClusterUserManager.SYSTEM_DB;
