+++ test/behaviour/connection/ConnectionStepsBase.java
@@ -24,7 +24,7 @@ package com.vaticle.typedb.client.test.behaviour.connection;
 import com.vaticle.typedb.client.api.connection.TypeDBClient;
 import com.vaticle.typedb.client.api.connection.TypeDBSession;
 import com.vaticle.typedb.client.api.connection.TypeDBTransaction;
-import com.vaticle.typedb.client.api.database.Database;
+import com.vaticle.typedb.client.api.connection.database.Database;
 import com.vaticle.typedb.common.test.server.TypeDBSingleton;
 
 import java.util.ArrayList;
